feat: Add msg_macro
This is a macro that creates simple messages without fields. This is the
minimal possible message suitable for using across binaries.

Also added to_serde_json_string as a compliment to from_serde_json_str/buf.

Change msg1 to just use msg_macro, but left msg1 as is so it can be a
development platform for future enhancement and debugging msg_marco.