feat: Add protocol_set
Now there are two entities a `Protocol` which has a name, id and a vector
of `MsgId`s. And then a `ProtocolSet` which is a set of `Protocol`s that
an actor supports.

Removed echo_protocol and now have EchoStartCompleteProtocol and
EchoReqReplyProtocol.  EchoStartCompleteProtocol has 2 messages EchoStart
and EchoComplete. And, as you'd guess, EchoReqReplyProtocol also happens to
have 2 messages EchoReq and EchoReply.

The Client has a ProtocolSet that has both Protocols where as the Server
only supports the EchoReqReplyProtocol.