refactor: Use AnId instead of the New Type
This simplifies the code, i.e. you see alot of deleted code! In the future
I may reintroduce New Types for some of these I think simpler is better
right now and it drives home that the primary type is an AnId and I need
to make it "universally unique" and that the type itself is not enough
to make it unique.

It has one other benefit, the underlying uuid is Copy which reduces the
need for `.clone()`.