refactor: move con_mgr_add_actor
Move it from msgs/con_mgr_add_actor to protocols/con_mgr_register_actor.
As it now defines not only ConMgrRegisterActorReq and ConMgrRegisterActorRsp
message but also ConMgrRegisterProtocol and ConMgrRegisteeProtocol.

Doing this means we consolidate the protocol and the messages in the
protocol into one package. This make maintenance easier although it
means we're discouraging using these messages in other protocols.

Also renamed ReqProtocolSet and RspProtocolSet to unused1 and 2 as we'll use
their AnId's in the future.