test: Add testing for 1 and 2 actors registering
Rename test-client to test_reg_client_server and have it first test
adding Client and then in that same test add Server so there are now two
actors registered. In particular we see that the echo_requestee_protocol
has both the Client and Server actor with the vec![0, 1] as expected.