feat: Adding notion of supervisor
ATM, actor_executor::test_con_mgr_server completes!

And actor_executor_protocol so ActorExecutors can eventually
be added to ConMgr's so thier instance_id's and rsp_tx be added
to RSP_TX_HASHMAP.

The problem is I have a chicken and egg problem, I create a ActorExecutor
and start it.  I then create a ConMgr and want to send AddActorReq, but I
need a src_id in the AddActorReq msg, but I don't have one.

ATM it seems "natural" to add the notion of a "supervisor" which is the
"main" thread so that's the direction I'm going. But another option might
be to use the instance_id of the ActorExecutor and pass that as the
supervisor.

Also, I'm adding so when ActorExecutor sends an AddActorRsp the
src_id needs an instance_id. And rather than just adding an instance_id
I thought it prudent to add actor_id and protocol_set too.

I will then be able to register ActorExecutors with con_mgr and they'll
be accessible via ConMgr queries and retrieving rsp_tx via
rsp_tx_map_by_instance_id.